requirements.txt
streamlit==1.31.0
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.0
tensorflow==2.17.0
keras==3.2.1
